+++ src/sampdf.jl
@@ -1,20 +1,22 @@
 """ 
-    sampdf(Y::DataFrame, id = 1:nro(Y); k, 
-        sampm = :rand)
-Build training/test sets for each column of a dataframe 
-    (typically, response variables to predict) that can contain missing 
-    values
+    sampdf(Y::DataFrame, k::Union{Int, Vector{Int}}, 
+        id = 1:nro(Y); msamp = :rand)
+Build training vs. test sets from each column of a dataframe. 
 * `Y` : DataFrame (n, p) whose each column can contain missing values.
 * `id` : Vector (n) of IDs.
 * `k` : Nb. of test observations selected for each `Y` column. 
     The selection is done within the non-missing observations 
     of the considered column. If `k` is a single value, the same nb.  
-    observations are selected for each column. Alternatively, `k` can 
-    be a vector of length p. 
-* `sampm` : Type of sampling for the test set.
-    Possible values are: :rand (default) = random sampling, 
-    :sys = systematic sampling over each sorted `Y` column
-    (see function `sampsys`).  
+    of observations are selected for each column. Alternatively, 
+    `k` can be a vector of length p. 
+Keyword arguments:
+* `msamp` : Type of sampling for the test set.
+    Possible values are: :rand = random sampling, 
+    :sys = systematic sampling over each sorted 
+    `Y` column (see function `sampsys`).  
+
+Typically, dataframe `Y` contains a set of response variables 
+to predict.
 
 ## Examples
 ```julia
@@ -23,15 +25,21 @@ using DataFrames
 Y = hcat([rand(5); missing; rand(6)],
    [rand(2); missing; missing; rand(7); missing])
 Y = DataFrame(Y, :auto)
+n = nro(Y)
 
-sampdf(Y; k = 3)
-
-sampdf(Y; k = 3, sampm = :sys)
+k = 3
+res = sampdf(Y, k) 
+#res = sampdf(Y, k, string.(1:n))
+pnames(res)
+res.nam
+length(res.test)
+res.train
+res.test
 
 ## Replicated splitting Train/Test
 rep = 10
 k = 3
-ids = [sampdf(Y[:, namy]; k = k) for i = 1:rep]
+ids = [sampdf(Y, k) for i = 1:rep]
 length(ids)
 i = 1    # replication
 ids[i]
@@ -44,8 +52,8 @@ ids[i].nam[j]
 ```
 """
 function sampdf(Y::DataFrame, k::Union{Int, Vector{Int}}, 
-        id = 1:nro(Y); sampm = :rand)
-    @assert in([:rand; :sys])(sampm) "Wrong value for argument 'sampm'."
+        id = 1:nro(Y); msamp = :rand)
+    @assert in([:rand; :sys])(msamp) "Wrong value for argument 'msamp'."
     p = nco(Y)
     nam = names(Y)
     train = list(Vector, p)  
@@ -54,7 +62,7 @@ function sampdf(Y::DataFrame, k::Union{Int, Vector{Int}},
     @inbounds for i = 1:p
         y = Y[:, nam[i]]
         s_all = findall(ismissing.(y) .== 0)
-        if sampm == :rand   
+        if msamp == :rand   
             n = length(s_all)
             res = samprand(n, k[i])
         else